CmsKit - Refactoring & Page lookup improvement
