+++ modules/cms-kit/src/Volo.CmsKit.Admin.Application.Contracts/Volo/CmsKit/Admin/Menus/MenuDto.cs
@@ -7,6 +7,7 @@ using Volo.Abp.Application.Dtos;
 
 namespace Volo.CmsKit.Admin.Menus
 {
+    [Serializable]
     public class MenuDto : EntityDto<Guid>
     {
         public string Name { get; set; }
