+++ modules/cms-kit/src/Volo.CmsKit.Admin.Application.Contracts/Volo/CmsKit/Admin/Menus/MenuItemCreateInput.cs
@@ -3,6 +3,7 @@ using System.ComponentModel.DataAnnotations;
 
 namespace Volo.CmsKit.Admin.Menus
 {
+    [Serializable]
     public class MenuItemCreateInput
     {
         public Guid MenuId { get; set; }
