+++ modules/cms-kit/src/Volo.CmsKit.Admin.Application.Contracts/Volo/CmsKit/Admin/Menus/PageLookupDto.cs
@@ -1,7 +1,7 @@
 using System;
 using Volo.Abp.Application.Dtos;
 
-namespace Volo.CmsKit.Admin.Pages
+namespace Volo.CmsKit.Admin.Menus
 {
     public class PageLookupDto : EntityDto<Guid>
     {
