+++ modules/cms-kit/src/Volo.CmsKit.Admin.Application.Contracts/Volo/CmsKit/Admin/Menus/PageLookupInputDto.cs
@@ -1,0 +1,11 @@
+using System;
+using Volo.Abp.Application.Dtos;
+
+namespace Volo.CmsKit.Admin.Menus
+{
+    [Serializable]
+    public class PageLookupInputDto: PagedAndSortedResultRequestDto
+    {
+        public string Filter { get; set; }
+    }
+}
