+++ modules/cms-kit/src/Volo.CmsKit.Admin.Application.Contracts/Volo/CmsKit/Admin/Menus/UpdateMainMenuInput.cs
@@ -1,5 +1,8 @@
+using System;
+
 namespace Volo.CmsKit.Admin.Menus
 {
+    [Serializable]
     public class UpdateMainMenuInput
     {
         public bool IsMainMenu { get; set; }
