+++ modules/cms-kit/src/Volo.CmsKit.Admin.Application.Contracts/Volo/CmsKit/Admin/Pages/IPageAdminAppService.cs
@@ -8,6 +8,5 @@ namespace Volo.CmsKit.Admin.Pages
 {
     public interface IPageAdminAppService : ICrudAppService<PageDto, PageDto, Guid, GetPagesInputDto, CreatePageInputDto, UpdatePageInputDto>
     {
-        Task<PagedResultDto<PageLookupDto>> GetLookupAsync(GetPagesInputDto input);
     }
 }
