+++ modules/cms-kit/src/Volo.CmsKit.Admin.HttpApi/Volo/CmsKit/Admin/Menus/MenuAdminController.cs
@@ -119,5 +119,13 @@ namespace Volo.CmsKit.Admin.Menus
         {
             return MenuAdminAppService.UpdateMenuItemAsync(menuId, menuItemId, input);
         }
+        
+
+        [HttpGet]
+        [Route("lookup/pages")]
+        public virtual Task<PagedResultDto<PageLookupDto>> GetPageLookupAsync([FromQuery]PageLookupInputDto input)
+        {
+            return MenuAdminAppService.GetPageLookupAsync(input);
+        }
     }
 }
